Fix memory-safety issues found by ASAN/UBSan
- Teardown leak (~42MB): ws_server_destroy never freed per-thread resources
  (ws_server_stop_internal joined threads but didn't call ws_cleanup_event_thread,
  and cleanup never destroyed the MPSC queue). Destroy now cleans each thread and
  ws_cleanup_event_thread destroys the message queue.
- Undefined behavior in frame handling (untrusted input):
  - ws_utils.c SHA-1: cast bytes to uint32_t before <<24 (signed overflow).
  - ws_frame.c unmask: memcpy instead of an unaligned uint32_t store (UB; crashes
    on strict-alignment archs).

+++ src/ws/ws_utils.c
@@ -98,10 +98,10 @@ int ws_sha1(const uint8_t *data, size_t len, uint8_t *hash) {
         
         /* Break chunk into words */
         for (int i = 0; i < 16; i++) {
-            w[i] = (msg[chunk + i * 4] << 24) |
-                   (msg[chunk + i * 4 + 1] << 16) |
-                   (msg[chunk + i * 4 + 2] << 8) |
-                   (msg[chunk + i * 4 + 3]);
+            w[i] = ((uint32_t)msg[chunk + i * 4] << 24) |
+                   ((uint32_t)msg[chunk + i * 4 + 1] << 16) |
+                   ((uint32_t)msg[chunk + i * 4 + 2] << 8) |
+                   ((uint32_t)msg[chunk + i * 4 + 3]);
         }
         
         /* Extend words */
